% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnalyzeMetaboliteData.R
\name{Num.Instruments}
\alias{Num.Instruments}
\title{Choose the number of potential instruments
Choose the number of potential instruments using a q-value threshold}
\usage{
Num.Instruments(Y, Cov = NULL, max.miss.C = 0.05,
  max.missing.consider = 0.5, K.max = NULL, q.thresh = c(0.01, 0.05,
  0.1))
}
\description{
Choose the number of potential instruments
Choose the number of potential instruments using a q-value threshold
}
